Major Fix to Music. Various Updates.
So after investigation I found that the executors used to download the Music Files, were staying alive and causing massive memory leaks. This new commit addresses that issue and fixes all known memory leaks.

This fix has come with a slight downfall, since we rely on a non-asynchronous callback, UI has been harder to implement. It's not impossible by any means but it's not something I will be working on in the near future. This means that when requesting a song, you will no longer see which song you have requested. Or the old Download ProgressBar.

Apart from that, performance wise, the music player should be performing much more effeciently.
Though it should be noted there are bound to be heaps of UI issues. Which is something which will need to be fined tuned as time goes.

NASA Commands have been added.
DOfus Commands have been added.

Fixed a small bug in the paginator, causing 2 pages to spawn randomly sometimes.

Handler now creates a gist if the error message is over 2024 characters in length.

Small fix to statistics, though this is bound to break, it's really a trial and error thing.

Added ffmpeg files for Windows.